+++ src/main/kotlin/net/irisshaders/lilybot/LilyBot.kt
@@ -18,13 +18,12 @@ import net.irisshaders.lilybot.extensions.events.JoinLeaveEvent
 import net.irisshaders.lilybot.extensions.events.MessageEvents
 import net.irisshaders.lilybot.extensions.moderation.Moderation
 import net.irisshaders.lilybot.extensions.moderation.Report
-import net.irisshaders.lilybot.extensions.moderation.ThreadModInviter
-import net.irisshaders.lilybot.extensions.support.ThreadInviter
 import net.irisshaders.lilybot.extensions.util.CustomCommands
 import net.irisshaders.lilybot.extensions.util.Github
-import net.irisshaders.lilybot.extensions.util.Ping
 import net.irisshaders.lilybot.extensions.util.RoleMenu
 import net.irisshaders.lilybot.extensions.util.ThreadControl
+import net.irisshaders.lilybot.extensions.util.ThreadInviter
+import net.irisshaders.lilybot.extensions.util.Utilities
 import net.irisshaders.lilybot.utils.BOT_TOKEN
 import net.irisshaders.lilybot.utils.CUSTOM_COMMANDS_PATH
 import net.irisshaders.lilybot.utils.GITHUB_OAUTH
@@ -71,7 +70,6 @@ suspend fun main() {
 
 		extensions {
 			add(::Config)
-			add(::Ping)
 			add(::Moderation)
 			add(::ThreadInviter)
 			add(::Report)
@@ -79,9 +77,9 @@ suspend fun main() {
 			add(::MessageEvents)
 			add(::Github)
 			add(::CustomCommands)
-			add(::ThreadModInviter)
 			add(::ThreadControl)
 			add(::RoleMenu)
+			add(::Utilities)
 
 			extPhishing {
 				appName = "Lily Bot"
