+++ src/main/kotlin/net/irisshaders/lilybot/extensions/util/Utilities.kt
@@ -4,12 +4,15 @@ package net.irisshaders.lilybot.extensions.util
 
 import com.kotlindiscord.kord.extensions.DISCORD_BLACK
 import com.kotlindiscord.kord.extensions.DISCORD_BLURPLE
+import com.kotlindiscord.kord.extensions.DISCORD_GREEN
+import com.kotlindiscord.kord.extensions.DISCORD_YELLOW
 import com.kotlindiscord.kord.extensions.checks.hasPermission
 import com.kotlindiscord.kord.extensions.commands.Arguments
 import com.kotlindiscord.kord.extensions.commands.converters.impl.boolean
 import com.kotlindiscord.kord.extensions.commands.converters.impl.string
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.ephemeralSlashCommand
+import com.kotlindiscord.kord.extensions.extensions.publicSlashCommand
 import com.kotlindiscord.kord.extensions.types.respond
 import dev.kord.common.entity.Permission
 import dev.kord.common.entity.PresenceStatus
@@ -17,9 +20,12 @@ import dev.kord.common.entity.Snowflake
 import dev.kord.core.behavior.channel.GuildMessageChannelBehavior
 import dev.kord.core.behavior.channel.createEmbed
 import dev.kord.core.behavior.channel.createMessage
+import dev.kord.rest.builder.message.create.embed
 import kotlinx.datetime.Clock
 import net.irisshaders.lilybot.database.DatabaseManager
 import net.irisshaders.lilybot.utils.ResponseHelper
+import net.irisshaders.lilybot.utils.TEST_GUILD_CHANNEL
+import net.irisshaders.lilybot.utils.TEST_GUILD_ID
 import org.jetbrains.exposed.sql.select
 import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
 import kotlin.time.ExperimentalTime
@@ -29,6 +35,43 @@ class Utilities : Extension() {
 	override val name = "utilities"
 
 	override suspend fun setup() {
+		val onlineLog = kord.getGuild(TEST_GUILD_ID)?.getChannel(TEST_GUILD_CHANNEL) as GuildMessageChannelBehavior
+
+		/**
+		 * Online notification
+		 * @author IMS212
+		 */
+		ResponseHelper.responseEmbedInChannel(onlineLog, "Lily is now online!", null, DISCORD_GREEN, null)
+
+		/**
+		 * Ping Command
+		 * @author IMS212
+		 * @author NoComment1105
+		 */
+		publicSlashCommand {  // Public slash commands have public responses
+			name = "ping"
+			description = "Am I alive?"
+
+			action {
+				val averagePing = this@Utilities.kord.gateway.averagePing
+
+				respond {
+					embed {
+						color = DISCORD_YELLOW
+						title = "Pong!"
+
+						timestamp = Clock.System.now()
+
+						field {
+							name = "Your Ping with Lily is:"
+							value = "**$averagePing**"
+							inline = true
+						}
+					}
+				}
+			}
+		}
+
 		/**
 		 * Say Command
 		 * @author NoComment1105
@@ -37,7 +80,7 @@ class Utilities : Extension() {
 			name = "say"
 			description = "Say something through Lily."
 
-			check { hasPermission(Permission.ModerateMembers) } // Idk wasn't sure
+			check { hasPermission(Permission.ModerateMembers) } // Wasn't sure about this
 
 			action {
 				var actionLogId: String? = null
